Simple extract with problems
